datj/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
mysqlclient @ file:///D:/Documents/PythonProjects/DatJ/mysqlclient-1.4.6-cp38-cp38-win32.whl
product==0.1.1.1
pytz==2020.1
sqlparse==0.4.1
